Add Scope#scope method
The `Model` class has a `scope()` method to return the scope for it.
It would be nice to be able to pass in either a model or a scope to the
a function and use the `scope()` method to cast it to a scope
regardless, instead of doing type detection.