Replace raw result array with CollectionProxy
This wraps the results of any collection requests in an CollectionProxy
instance, which allows users to access additional metadata about the
result set. Users can call `CollectionProxy.data` to get the parsed model
array, or may access the original request body with `CollectionProxy.raw`.
Additionally, the `meta` field of the response is exposed as a
first-class member, in `CollectionProxy#meta`.

Follow-up items include similar wrapping for scope and model methods
that are expecting a single result, wrapping them in a `RecordProxy` or
something similar.