Fix bug with deeply nested associations
This fixes an issue when looking up/assigning a deep assocation that had
already been assigned to another object would assign an empty copy of
the object instead of the target.  It was doing this because the return
value of the lookup wasn't being used, and instead the function was
inadvertently relying on the modification to the passed in parameter.