Make error validation immutable
Previously the model errors object was being cleared out on save and
then the validator was appending all errors to that empty object after
the result set came back.  This is problematic for frameworks like
vue.js which rely on object observers, as an observer change event is
not fired when the new errors are added unless the developer
specifically requests a deep object watch, with is both computationally
expensive and something that has to be done every time.  This replaces
the behavior by creating a new errors object each time the validator is
run and replacing the one currently on the model with it.