Move to predictive temp IDs
Temp IDs don't have to actually be unique - they just have to match up
the relationships/included sections of the payload, and serve as a
reference to an unpersisted object on the client. They should be unique,
not universally or globally unique.

Because of this, let's generate more predictive, simpler temp IDs. This
improves readability, and avoids mocking difficulties.