Add strictAttributes
As part of the initial 1.0 refactor I'd made the object initialization
throw an error for any unknown attributes.  This was incompatible with
explicit tests to the contrary in the old test suite, and I realized it
obviously isn't ideal for when the underlying response payload includes
fields not defined by the model.  I figure this is still desired for
some use-cases though, so I've added a `strictAttributes` static
property to model classes that will throw an error if set to true.