Allow decorators to be called without @syntax
Decorators are still a fairly new feature of JS and are not yet widely
supported outside of typescript and cutting-edge Babel. Needed to do
some cleanup to the decorators anyway and realized it would be simple
to make calling the decorator functions to declare attributes and
associations on ES6 classes directly.  This is probably as far as we'll
go in supporting raw ES6 classes for the time being, but this lets
anyone who really wants to use them have a path for that.