Greatly simplify type definitions for Model
Previously, the type definitions needed to get the `JSORMBase.extend()`
API working with correct typings was complicated to say the very least.
It's a lot simpler now, though with a caveat that we can't use
typescript classes to extend a class that is created from the
`extend()` API, which seems like a worthwhile limitation for the
tradeoff.

Also adds a types test that simply validates that both the class and
extend APIs both provide the same interface.