Ensure singular calls always resolve a record
Previously the `response.data` of a `Person.find(1).then((response))`
call could be a `Person` instance or null. This is valid according to
the spec, but not helpful for actual development.

Instead, we reject the promise with a relevant error message if the
response is 404 or if `data` is `null`. This way we can have confidence
a successful promise always has not-null `data`