Avoid blowing up on valid #destroy() response
According to http://jsonapi.org/format/#crud-deleting, we should not
expect to get a resource back from a DELETE request. Instead we should
expect 202/204 to render no content, and 200 (the new default in
jsonapi_suite) to render only a top-level `meta`.

This code may need to be refactored if we needed to parse meta or fire
afterFetch middleware, but I think we can safely punt on that for now.