Add inferred type warning in dev mode  (#55)
* Add inferred type warning in dev mode

When using the `@Model` decorator API, the jsonapi type will be inferred
based on the class name. This is really convenient for getting up and
running quickly, but when building code into a minified browser bundle,
the class names are minified as well, ruining the inferrence.  This
commit detects if the user is running in non-production mode in a
browser and provides a console warning if the inferrence code is called.

* build bundle