Move the Omit utility type to a .ts file (#56)
* Move the Omit utility type to a .ts file

Ran into an issue with a consuming application that had a config
that tried to validate the types of dependent libraries. In this case
the types were failing because things from `types/index.d.ts` weren't
available externally. Moving them to actual module with exports fixes
it.  Shouldn't be necessary in most cases, but fixes the problems for a
few edge-case configurations.

* Add build