Create Error class similar to ActiveModel
Previously a model's errors was simply a hash of strings.  This made it
difficult to do things like getting the "full message" including a
humanized attribute name, or access the underlying error code.  This
replaces the previous version with an actual object that has all of
these things available.  Additionally, the error object is typed so
that it only includes the attributes available on the model in question.