Avoid syncing dirty props on store update
This is probably a better default, following principle of least
surprise, unless we hear otherwise. If you load instanceA, and 1m later
load instanceB with new data from the server, instanceA will change. But
if instanceA has already updated the relevant attribute (it is dirty),
avoid making that change.

Relatedly, if instanceA syncs with a new store update, it would see that
property as dirty, as it didn't match the attributes it was instantiated
with. This commit also ensures synced properties are not seen as dirty.