Revert switch to TS 2.8's `Exclude` (#96)
`Exclude` is a new language feature from typescript 2.8 that replaces the
`Diff` functionality previously backing the `Omit` type.  This only
works in clients using 2.8, so let's keep using the old version for a
little while until we can be confident most people have upgraded. Or
until we can figure out how to have some sort of conditional typecheck
in typescript.