Get working on all TS version >= 2.8.0
This gets everything working on newer versions of typescript, with the
unfortunate tradeoff of removing support of all versions before 2.8.
Due to some changes in 2.9 we need to take advantage of some type
extraction utilities that are introduced in version 2.8 to make all of
our generics compatible.  Seems worth it though since we get everything
working with modern TS and this is part of a renamed library with a
separate major versioning scheme.

Also adds the `yarn test:all-versions` which will test against every
published version of typescript in the supported range.