Send Client-Application header in requests
If you have this set:

```ts
class ApplicationRecord extends SpraypaintBase {
  static clientApplication = 'contract-detail'
}
```

All requests will send the header:

```
Client-Application: contract-detail
```

This is helpful for servers to record which clients are consumers of the
API.

Notably, this is a simple static property. If you define all your
models, then set `ApplicationRecord.clientApplication = 'foo'` afterwards, it
will not be picked up (I do not believe accomodating this warrants the
additional complexity).