Convert from TSLint to ESLint
tslint was first setup around a year ago, but ended up being a pretty
big pain to deal with, so it was turned off as an automatic thing and
the rules were no longer strictly followed.  Now that eslint has a
robust typescript parser, we no longer need to rely on tslint and can
take advantage of a better ecosystem with more sane rules. This commit
switches to that and pulls in recommended rulesets for eslint and
prettier

Also adds a commit hook to automatically run staged files through an
autofixer on commit.