Scope<> generic uses instance instead of class
Previously the return type of `Author.scope()` was
`Scope<typeof Author>`.  For typing purposes this made it easy to get
both the class and instance types using `ClassType['prototype']`.  This
had a limitation though where we could not instantiate an
explicitly-typed `Scope` from an instance method, as typescript doesn't
provide an easy way to go from `Type` to `typeof Type` in generic
declarations, which is generally frustrating.  This changes the
generics definition of `Scope`, though I haven't seen any use of these
types directly in the codebases I've worked with, so I'm not super
concerned about the types breaking (and it should be an easy upgrade),
expecially since we haven't released Spraypaint 1.0 yet.

The only "loss" in this change is that the `model` member within `Scope`
is now not a generic type and directly typed to `typeof SpraypaintBase`,
but since we aren't really relying on anything about this generic, I'm
not worried about losing out on it.