Get tc39 decorators working
When the decorators were first setup, typescript and Babel had the same
API for implementing decorators.  Since then, TC39 has evolved the
decorator API.  This adds compatibility with the newer decorators.
Additionally, it adds an es6 compatibility smoke suite which will run
after the ESM module build.

The decorators code is getting pretty hairy with a number of code paths,
so I'd like to clean it up, but this at least gets it working an
unblocks a few users who are working in a modern babel setup.