Allow Number types not to write empty string
Previously there was an issue when binding a Spraypaint model to a form
or otherwise allowing a number field to be set to an empty string.  This
would pass down to the backend and Graphiti would (correctly) blow up,
as it was receiving a string but expecting a number (or null, in this
case).  This will obey specyfing `Attr({type: Number})` in the model
definition.  When this receives an empty string during payload
serialization, it will write `null` instead.

This is a quick fix to a specific user problem.  There is a bigger
commit that needs to happen here with casting & payload
serialization/deserialization, but I'd like to get this in to fix the
immediate concern for now.